edit item improved

+++ app/controllers/index.js
@@ -8,25 +8,29 @@ export default class IndexController extends Controller {
 
   @tracked showAddModal = false;
   @tracked selectedItem;
-
+  @tracked selectedIndex;
   get items() {
     return this.model.filter((item) => !item.isDeleted);
   }
 
   @action
   submitNewItem() {
-    if (!this.selectedItem.id) {
-      this.selectedItem.id = crypto.randomUUID();
+    if (this.selectedIndex !== null) {
+      this.items[this.selectedIndex].setProperties(this.selectedItem);
+    } else {
       this.store.createRecord('item', this.selectedItem);
     }
     this.showAddModal = false;
     this.selectedItem = null;
+    this.selectedIndex = null;
     return;
   }
 
   @action
   createNewItem() {
+    this.selectedIndex = null;
     this.selectedItem = {
+      id: crypto.randomUUID(),
       name: '',
       value: 0,
       category: '',
@@ -61,12 +65,21 @@ export default class IndexController extends Controller {
 
   @action
   deleteItem(item) {
-    item.deleteRecord();
+    if (confirm('Are you sure you want to delete this item?')) {
+      item.deleteRecord();
+    }
   }
 
   @action
-  editItem(item) {
-    this.selectedItem = item;
+  editItem(item, index) {
+    this.selectedItem = {
+      id: item.id,
+      name: item.name,
+      value: item.value,
+      category: item.category,
+      itemDetail: item.itemDetail,
+    };
+    this.selectedIndex = index;
     this.showAddModal = true;
   }
 
@@ -74,10 +87,11 @@ export default class IndexController extends Controller {
   cancelItem() {
     this.showAddModal = false;
     this.selectedItem = null;
+    this.selectedIndex = null;
   }
 
   get total() {
-    let totalValue=0;
+    let totalValue = 0;
     this.items.forEach((item) => {
       const value = parseFloat(item.value) || 0;
       totalValue += value;
